Remove cordovaDependencies - does not work with pre-release versions (#183)
Installation on VS fails because the cordovaDependencies resolution logic doesn't work with pre-release versions. Note that removing it will simply result in installation defaulting to the latest version. This should be acceptable, as it is highly unlikely at this point for a new user to be using a version of Cordova that we don't support, and even if they are, there is no version of our plugin that they would be able to use (and our docs do document this limitation clearly).